Refactor radio item fields to use a generic radio group component
